refactor(frontend): Modificamos el flujo para carga de menú y acceso a la aplicación.
