+++ apps/frontend/src/app/app-routing.module.ts
@@ -4,13 +4,14 @@ import { AppHomeComponent } from './home/home.component';
 import { RoutingNavBar, RoutingGuard } from './login/routing-guard';
 
 const appRoutes: Routes = [
+
+    { path: 'auth', loadChildren: './login/login.module#LoginModule', canActivate: [RoutingNavBar] },
     {
-        path: '',
+        path: 'home',
         component: AppHomeComponent,
         canActivate: [RoutingGuard, RoutingNavBar],
         pathMatch: 'full'
     },
-    { path: 'auth', loadChildren: './login/login.module#LoginModule', canActivate: [RoutingNavBar] },
     {
         path: 'institution',
         loadChildren: './institutions/institution.module#InstitutionModule',
@@ -27,7 +28,7 @@ const appRoutes: Routes = [
         canActivate: [RoutingNavBar, RoutingGuard]
     },
 
-    { path: '**', redirectTo: '/' }
+    { path: '**', redirectTo: '/home' }
 ];
 
 export const AppRouting: ModuleWithProviders = RouterModule.forRoot(appRoutes);
