+++ apps/frontend/src/app/home/home.routing.ts
@@ -1,0 +1,10 @@
+import { Routes, RouterModule } from '@angular/router';
+import { NgModule } from '@angular/core';
+
+const routes: Routes = [];
+
+@NgModule({
+    imports: [RouterModule.forChild(routes)],
+    providers: []
+})
+export class InstitutionRouting { }
